simplify the use of build tags and config
