+++ server/config.go
@@ -1,5 +1,3 @@
-//+build !lambda
-
 package server
 
 import (
