--- startup/startup.go
@@ -1,9 +1,0 @@
-//+build !lambda
-
-package startup
-
-import "github.com/joshprzybyszewski/cribbage/server"
-
-func PlayServer() error {
-	return server.Setup()
-}
